re-wrote serialization for the watch list config
The extensions on object wasn't gelling well, moved it to a more obvious
SerializationHelper. Also removed the need to be able to get a
serializer for oneself which made the whole thing cleaner.

Also added the console sandbox (default start project for now) to launch
the xml file that it generates and also to dump the path to it in the
debug console. It goes for the temp directory. (*hope I didn't hard-code
c:\temp")